+++ crates/automod/src/storage/mod.rs
@@ -1,8 +1,0 @@
-//! Contains modules of operations on redis used by the automod
-
-pub mod allow_list;
-pub mod config;
-pub mod history;
-pub mod lock;
-pub mod playlist;
-pub mod speaker;